+++ www/static/global/js/main.js
@@ -222,11 +222,9 @@ function SetReload(time){
 }
 /* ReloadPage -------------------------------------- */
 function ReloadPage(jump) {
-	var reload_album=$('#jsReload').attr('data-pf_album');
-	var reload_url	=$('#jsReload').attr('data-pf_url');
-	if(jump == 1 && reload_album){
-		var url=reload_url+reload_album+'&back=1';
-		window.location.href=url;
+	var reload_url=$('#jsReload').attr('data-url');
+	if(jump == 1 && reload_url !=''){
+		window.location.href=reload_url;
 	}
 	else{
 		location.reload();